refactor(search): extract SearchFilter service from search command
Extracts applyFilters(), atomHasTrailer(), and atomMatchesText() into a
dedicated SearchFilter service class. Replaces the 12-case switch in
atomHasTrailer() with a data-driven approach using ARRAY_TRAILER_KEYS
and ENUM_TRAILER_KEYS constants. Removes unused LORE_TRAILER_KEYS import
from search.ts.
